Refactor main.py and remove redundant method.py
This commit refactors the main.py file to make code more modular and readability high. The changes include creation of functions for each specific task such as send_request, write_file and check_status_code (previously this was all lumped together) to improve code structure. The logic part of the original main.py is moved into a main function to make it easier to test. Furthermore, it modernizes the user inputs using the 'rich' library for better UI.

On the other hand, the method.py file is removed because it contained identical POST request methods as main.py and was causing redundancy in our codebase.